% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_songs.R
\docType{data}
\name{mood_songs}
\alias{mood_songs}
\title{Built-in Song Dataset}
\format{
A data.frame with 364 rows and feature columns:
energy, valence, danceability, tempo_norm, acousticness, loudness_norm
}
\usage{
mood_songs
}
\description{
A dataset of 364 songs with six audio features and a reference mood label.
This is the same dataset returned by load_songs().
}
\examples{
# Access built-in dataset
songs <- load_songs()
dim(songs)

}
\keyword{datasets}
