% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_songs.R
\name{print.mood_songs}
\alias{print.mood_songs}
\title{Print Method for mood_songs}
\usage{
\method{print}{mood_songs}(x, ...)
}
\arguments{
\item{x}{A mood_songs object}

\item{...}{Additional arguments (unused)}
}
\description{
Displays a quick summary of the dataset including number of songs,
features, mood distribution (if available), and scaling status.
}
\examples{
songs <- load_songs()
print(songs)

}
