% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VibeCluster-package.R
\docType{package}
\name{VibeCluster}
\alias{VibeCluster-package}
\alias{VibeCluster}
\title{VibeCluster: Mood-Based Music Clustering and Recommendation}
\description{
VibeCluster groups songs into mood clusters using k-means on audio features
(energy, valence, danceability, tempo, acousticness, loudness).
It also provides visualisation using PCA and song recommendations
using nearest neighbours.
}
\section{Methods Used}{

\itemize{
  \item K-means clustering for grouping songs
  \item PCA for 2D visualisation (Mood Map)
  \item Nearest neighbour search for recommendations
  \item  Simple k Selection using Within-Cluster Sum of Squares
}
}

\section{Workflow}{

\preformatted{
library(VibeCluster)

songs <- load_songs()
mood_summary(songs, k_range = 2:8)

result <- mood_cluster(songs, k = 6)
plot_mood_map(result)

get_mood_palette(result)

my_mood <- c(energy = 0.8, valence = 0.3, danceability = 0.6,
             tempo_norm = 0.75, acousticness = 0.1, loudness_norm = 0.85)

recommend_songs(result, profile = my_mood, n = 8)
}
}

\section{Data}{

The package includes a dataset of 364 songs with mood labels.
Use load_songs() to access it.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/example/VibeCluster}
  \item Report bugs at \url{https://github.com/example/VibeCluster/issues}
}

}
\author{
\strong{Maintainer}: Student Author \email{author@example.com}

}
